refactor(quiz): extract TopicList from topic-tree + post-commit test/doc fixups
Post-commit review fixups for the RT setup shell:

- code-reviewer (§1 same-commit extraction): the showCode threading worsened the
  already-over-cap topic-tree.tsx (161→171, cap 150). Extract SubtopicRows + TopicList
  into topic-list.tsx; topic-tree.tsx drops to 81. No behavior change (shared quiz
  component — same rendered output, existing tests pass unmodified).
- test-writer: cover showCode=false on topic-row; cover the isPending/authError
  StartButton-disabled states on the RT form.
- semantic-reviewer: add an RT topic-load assertion to the get-rt-subject integration
  test (the client-side fetchTopicsWithSubtopics path a user hits on mount).
- doc-updater: note in docs/plan.md that Phase 1's bespoke RT hooks were superseded
  by the shared-hook reuse in this slice (history preserved, forward pointer added).
